+++ NEWS.md
@@ -1,5 +1,7 @@
 # NMdata 0.2.5
 ## New Features
+* `reorder2()`
+
 * `setOrderList()` is a new convenient function that reorders elements in a list, much like `setcolorder()` for data.table's. It takes both `first` and `last` arguments to specify what elements to move where.
 
 * `dcastSe()` gains ability to rename left-hand-side columns directly.